Revert "Users will automatically be logged out after 60 minutes (#54)"
This reverts commit 19087a8fd9f3bf7306f24aea63579a4c69a3f666.
