+++ drive-stats/src/components/login.jsx
@@ -17,7 +17,36 @@ import {
     CardTitle,
 } from '@/components/ui/card'
 import { useNavigate } from 'react-router-dom'
-import { useTimer } from './timer' // Import TimerContext
+import axios from 'axios'
+
+export async function handleLogout(setUser, toast) {
+    const API_HOST = import.meta.env.VITE_API_HOST
+    try {
+        const response = await axios.delete(`${API_HOST}/api/auth/signout`, {
+            withCredentials: true,
+        })
+
+        if (response.status === 200) {
+            sessionStorage.removeItem('token')
+            sessionStorage.removeItem('username')
+            setUser(null)
+            toast({
+                title: 'Logged Out',
+                description: 'You have been logged out successfully.',
+            })
+        } else {
+            throw new Error('Failed to log out')
+        }
+    } catch (error) {
+        console.error('There was an error logging out!', error)
+        toast({
+            title: 'Error',
+            description:
+                'There was an error logging out. Please try again later.',
+            status: 'error',
+        })
+    }
+}
 
 function LoginForm({
     isSignupFormOpen,
@@ -30,13 +59,12 @@ function LoginForm({
     const [username, setUsername] = useState('')
     const [password, setPassword] = useState('')
     const [isLoading, setIsLoading] = useState(false)
-    const { setUser } = useContext(UserContext) // Get user context
+    const { setUser } = useContext(UserContext)
     const { toast } = useToast()
-    const navigate = useNavigate() // Ensure navigate is defined here
+    const navigate = useNavigate()
     const [error, setError] = useState('')
     const formRef = useRef(null)
     const API_HOST = import.meta.env.VITE_API_HOST
-    const { startLogoutTimer, clearLogoutTimer } = useTimer() // Use TimerContext
 
     async function handleLogin(e) {
         e.preventDefault()
@@ -69,7 +97,6 @@ function LoginForm({
                     title: 'Logged In',
                     description: `Welcome back, ${username}`,
                 })
-                startLogoutTimer() // Start the logout timer on successful login
             } else {
                 setError('Wrong username or password.')
             }
@@ -151,36 +178,20 @@ function LoginForm({
                                 </Button>
                             </div>
                             <div className="w-full md:w-auto">
-                                <Button
-                                    onClick={() => {
-                                        console.log(
-                                            'Manual logout button clicked...'
-                                        )
-                                        handleLogout(
-                                            setUser,
-                                            toast,
-                                            navigate,
-                                            logoutTimerRef
-                                        )
-                                    }}
-                                    className="w-full md:w-auto"
-                                >
-                                    Test Logout
-                                </Button>
+                                <div className="hidden md:block">
+                                    <p className="text-sm text-gray-500">
+                                        New to DriveStats?{' '}
+                                        <button
+                                            type="button"
+                                            className="text-purple-600 hover:underline focus:outline-none focus:ring"
+                                            onClick={toggleSignUpForm}
+                                        >
+                                            Sign up here
+                                        </button>
+                                    </p>
+                                </div>
                             </div>
                         </CardFooter>
-                        <div className="text-center">
-                            <p className="text-sm text-gray-500">
-                                New to DriveStats?{' '}
-                                <button
-                                    type="button"
-                                    className="text-purple-600 hover:underline focus:outline-none focus:ring"
-                                    onClick={toggleSignUpForm}
-                                >
-                                    Sign up here
-                                </button>
-                            </p>
-                        </div>
                     </form>
                 </CardContent>
             </div>
